+++ app/(route)/(login)/signup/sign-up-from.tsx
@@ -11,8 +11,8 @@ import { useRouter } from "next/navigation";
 import { useState } from "react";
 import pulse from "@/public/icons/pulse.svg";
 import Image from "next/image";
-import useCheckLoginStatus from "@/app/_hooks/useCheckLoginStatus";
-import useModal from "@/app/_hooks/useModal";
+import useCheckLoginStatus from "@/app/_hooks/use-check-login-status";
+import useModal from "@/app/_hooks/use-modal";
 import ConfirmModal from "@/app/_components/modals/_components/confirm-modal";
 
 interface FormValues {
